Extract MenuBar & StatusBar to their corresponding component folders
