+++ new-log-viewer/src/components/modals/SettingsModal/index.tsx
@@ -2,6 +2,8 @@ import Modal from "@mui/joy/Modal";
 
 import ConfigDialog from "./ConfigDialog";
 
+import "./index.css";
+
 
 /**
  *
