+++ new-log-viewer/src/utils/actions.ts
@@ -1,6 +1,13 @@
 import * as monaco from "monaco-editor/esm/vs/editor/editor.api.js";
 
+import {updateWindowUrlHashParams} from "../contexts/UrlContextProvider";
 import {Nullable} from "../typings/common";
+import {CONFIG_KEY} from "../typings/config";
+import {getConfig} from "./config";
+import {
+    getFirstItemNumInNextChunk,
+    getLastItemNumInPrevChunk,
+} from "./math";
 
 
 enum ACTION_NAME {
@@ -62,8 +69,38 @@ const EDITOR_ACTIONS : ActionType[] = [
 ];
 /* eslint-enable sort-keys */
 
+/**
+ *
+ * @param actionName
+ * @param logEventNum
+ * @param numEvents
+ */
+const handleAction = (actionName: ACTION_NAME, logEventNum: number, numEvents: number) => {
+    const pageSize = getConfig(CONFIG_KEY.PAGE_SIZE);
+    switch (actionName) {
+        case ACTION_NAME.FIRST_PAGE:
+            updateWindowUrlHashParams({logEventNum: 1});
+            break;
+        case ACTION_NAME.PREV_PAGE:
+            updateWindowUrlHashParams({
+                logEventNum: getLastItemNumInPrevChunk(logEventNum, pageSize),
+            });
+            break;
+        case ACTION_NAME.NEXT_PAGE:
+            updateWindowUrlHashParams({
+                logEventNum: getFirstItemNumInNextChunk(logEventNum, pageSize),
+            });
+            break;
+        case ACTION_NAME.LAST_PAGE:
+            updateWindowUrlHashParams({logEventNum: numEvents});
+            break;
+        default:
+            break;
+    }
+};
 export {
     ACTION_NAME,
     EDITOR_ACTIONS,
+    handleAction,
 };
 export type {ActionType};
